requirements.txt
requests
Flask[async]
psycopg2-binary
sqlalchemy
APScheduler
pyTelegramBotAPI
python-dotenv
redis
pika